modules/yolo-detect/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tensorboard>=2.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
protobuf<4.21.3  # https://github.com/ultralytics/yolov5/issues/8012
pandas>=1.1.4
seaborn>=0.11.0